+++ db/migrations/000073_intercity_orders.up.sql
@@ -5,7 +5,7 @@ CREATE TABLE IF NOT EXISTS intercity_orders (
   to_location VARCHAR(255) NOT NULL,
   trip_type ENUM('companions','parcel','solo') NOT NULL,
   comment TEXT,
-  price INT NOT NULL,
+  price BIGINT NOT NULL,
   contact_phone VARCHAR(32) NOT NULL,
   departure_date DATE NOT NULL,
   departure_time TIME NULL,
