+++ db/migrations/000094_rent_ad_pricing_flags.down.sql
@@ -1,5 +1,5 @@
 ALTER TABLE rent_ad
-    MODIFY price DECIMAL(10, 2) NOT NULL,
+    MODIFY price BIGINT NOT NULL,
     DROP COLUMN price_to,
     DROP COLUMN negotiable,
     DROP COLUMN hide_phone;
